# chore(perf): only use the regular expression for shorter input