@cesco69 2024-03-12
@mcollina Any plan for a new version? there are some "merged" performance improvement

https://github.com/fastify/fast-json-stringify/pull/691
https://github.com/fastify/fast-json-stringify/pull/689
https://github.com/fastify/fast-json-stringify/pull/688
https://github.com/fastify/fast-json-stringify/pull/686